pyproject.toml
[project]
name = "kafka-chatroom"
version = "0.1.0"
description = "Project for kafka chatroom"
authors = [
    {name = "scottjasontaiwan"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "kafka-python (>=2.2.10,<3.0.0)",
    "pymongo (>=4.13.0,<5.0.0)",
    "flask (>=3.1.1,<4.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "flask-socketio (>=5.5.1,<6.0.0)",
    "eventlet (>=0.40.0,<0.41.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "confluent-kafka (>=2.10.0,<3.0.0)",
    "pytz (>=2025.2,<2026.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
